+++ packages/all/index.ts
@@ -1,9 +1,5 @@
 const config = {
-	extends: [
-		'@eslint-sets/eslint-config-basic',
-		'@eslint-sets/eslint-config-react',
-		'@eslint-sets/eslint-config-vue'
-	]
+	extends: ['@eslint-sets/eslint-config-react', '@eslint-sets/eslint-config-vue']
 }
 
 export default config